Aggiunti alcuni long che avevo dimenticato, scoperto un bug molto importante:
BUG:
Quando si fa il login dell'applicazione, si chiude e si riavvia l'applicazione, si crea un utente "fantasma": si è già loggati, ma come utente null. Il bug peraltro non si manifesta sempre allo stesso modo

+++ rentx_app/RentIT/RentIT/RentIT/Services/Authentication/AuthenticationService.cs
@@ -72,7 +72,7 @@ namespace RentIT.Services.Authentication
             return logOutInfo.hasSucceded;
         }
 
-        public int GetCurrentUserId()
+        public long GetCurrentUserId()
         {
             return AppSettings.UserId;
         }
